Remove inaccurate information about Workflow Execution Timeout (#40)
I have removed a sentence referencing a 10-year timeout for Workflow Executions, which is inaccurate. 

That default 10-year timeout was likely correct at the time of writing, but [Maxim stated in a Slack thread today](https://temporalio.slack.com/archives/CTQU95E84/p1670669082276379) that the default Workflow Execution timeout was changed to infinity two years ago. The documentation to which the incorrect sentence links also confirms that the default value is infinity.